+++ packages/types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @templatical/types
 
+## 0.24.1
+
 ## 0.24.0
 
 ### Minor Changes